fix error handling and move interface
