+++ src/pages/api/v1/auth/validate-token.ts
@@ -1,5 +1,5 @@
-import { validateJsonWebToken } from "@/services/auth";
-import { ApiError } from "@/types/exceptions";
+import { verifyToken } from "@/services/jwt";
+import { ApiError, UnauthorizedError } from "@/types/exceptions";
 import { NextApiRequest, NextApiResponse } from "next";
 
 export default async function handler(
@@ -13,8 +13,13 @@ export default async function handler(
   }
 
   try {
-    const response = await validateJsonWebToken(authorization as string);
-    res.status(200).json(response);
+    if (!authorization || !authorization.startsWith("Bearer ")) {
+      throw new UnauthorizedError("No token provided");
+    }
+    const token = authorization.split(" ")[1];
+
+    const decodedToken = verifyToken(token);
+    res.status(200).json({ isValid: true, decodedToken: decodedToken });
   } catch (error) {
     if (error instanceof ApiError) {
       res.status(error.statusCode).json({ error: error.message });
