+++ src/services/auth.ts
@@ -2,8 +2,6 @@ import {
   AlreadyExistsError,
   BadRequestError,
   DoesNotExistError,
-  InternalServerError,
-  UnauthorizedError,
 } from "@/types/exceptions";
 import {
   passwordsAreEqual,
@@ -12,7 +10,6 @@ import {
   validatePassword,
 } from "@/utils/auth";
 import bcrypt from "bcrypt";
-import jwt, { JsonWebTokenError } from "jsonwebtoken";
 import { createUser, findUserByEmail } from "../db/actions/auth";
 import { User } from "../db/models";
 
@@ -109,23 +106,4 @@ export async function validateLogin(email: string, password: string) {
   );
 
   return { token };
-}
-
-export async function validateJsonWebToken(authorization: string) {
-  if (!authorization || !authorization.startsWith("Bearer ")) {
-    throw new UnauthorizedError("No token provided");
-  }
-
-  const token = authorization.split(" ")[1];
-
-  try {
-    const decodedToken = jwt.verify(token, JWT_SECRET);
-    return { isValid: true, decodedToken: decodedToken };
-  } catch (error) {
-    if (error instanceof JsonWebTokenError) {
-      throw new UnauthorizedError("Invalid or expired token.");
-    } else {
-      throw new InternalServerError("An unknown error occurred.");
-    }
-  }
 }
